Split successful, not sure if the parse tree generated is in the right order though
