+++ Cargo.toml
@@ -1,14 +1,9 @@
-[package]
-name = "ppparser"
-version = "0.4.0"
-edition = "2021"
+[workspace]
 
-[build]
-profiler = true
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-[dependencies]
-
+members = [
+    "parser",
+    "generator"
+]
 
 [profile.release]
 strip = true  # Automatically strip symbols from the binary.
