+++ parser/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "parser"
+version = "0.4.0"
+edition = "2021"
+
+[build]
+profiler = true
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[dependencies]
+
+
