+++ parser/src/main.rs
@@ -1,11 +1,11 @@
-use ppparser::parser::Grammar;
-use ppparser::parser::Resolvable;
-use ppparser::utils::{embed_core, read_grammar, write_parser};
+use parser::parser::Grammar;
+use parser::parser::Resolvable;
+use parser::utils::{embed_core, read_grammar, write_parser};
 use std::env::args;
 use std::fs;
 use std::process::ExitCode;
-use ppparser::output_stack::Stack;
-use ppparser::cache::Cache;
+use parser::output_stack::Stack;
+use parser::cache::Cache;
 
 fn amain() {
     //let args: Vec<String> = env::args().collect().;
@@ -42,7 +42,7 @@ fn amain() {
     let core_parser: String = embed_core();
     let grammar = read_grammar(grammar);
     //  Temp ignore unused code with _grammmar since it's needed on line 39 later
-    let _grammar = match grammar {
+    let grammar = match grammar {
         Ok(grammar) => grammar,
         Err(_) => {
             println!(
@@ -52,16 +52,16 @@ fn amain() {
             std::process::exit(3);
         }
     };
-    let _grammar = _grammar + "\0";
-    let _grammar_length = _grammar.len() as u32 + 1;
+    let grammar = grammar + "\0";
+    let grammar_length = grammar.len() as u32 + 1;
     let position = 0;
-    let mut cache = Cache::new(_grammar_length, 43);
-    let mut stack = Stack::new(100,100);
+    let mut cache = Cache::new(grammar_length, 43);
+    let mut stack = Stack::new(grammar_length,43);
 
-    let (_bool, _position) = Grammar.resolve(&mut stack, &mut cache, position, &_grammar);
+    let (_bool, _position) = Grammar.resolve(&mut stack, &mut cache, position, &grammar);
     
 
-    stack.print_with_strings(&_grammar);
+    stack.print_with_strings(&grammar);
     // Add the parser generation here to then add into parser write
     let parser_write = write_parser(dest, &core_parser);
     match parser_write {
