+++ parser/src/output_stack.rs
@@ -1,9 +1,13 @@
 //
 
-use std::array::IntoIter;
+//
+
+use std::process::exit;
+
+use crate::parser::Rules;
 
 pub struct StackEntry{
-    pub rule: u32,
+    pub rule: Rules,
     pub start_position: u32,
     pub end_position: u32,
 }
@@ -13,7 +17,7 @@ pub struct Stack{
 }
 impl Stack{
     pub fn new(size_of_source: u32, number_of_structs: u32) -> Self {
-        let size_of_stack: usize = ((size_of_source*number_of_structs)/10).try_into().unwrap(); // /10 is a heuristic 
+        let size_of_stack: usize = ((3*size_of_source*number_of_structs)).try_into().unwrap(); // /10 is a heuristic 
         // Heuristic should be modified to match grammar profiling since some will use more stack and some less depending on language complexity.
         // TODO: Fundamentally Vec can grow so still need to catch unwrap above but it's merely about minimizing allocation rather than causing a break
         return Stack { entries: Vec::with_capacity(size_of_stack)};
@@ -31,17 +35,32 @@ impl Stack{
         return entry;
     }
 
-    pub fn print_with_strings(&self, source: &str){
+    pub fn print_with_strings(&self, source: &str)
+    {
         let stack_iterator = self.into_iter();
+        let mut print_vec: Vec<String> = Vec::new();
+        // Invert order because terminal appears inverted because of push down
         for entry in stack_iterator{
-            println!("{}, {}, {}, {}", entry.rule, entry.start_position, entry.end_position, &source[entry.start_position as usize..entry.end_position as usize])
+            let s: String = format!("{:?}, {}, {}, {}", entry.rule, entry.start_position, entry.end_position, &source[entry.start_position as usize..entry.end_position as usize]);
+            print_vec.push(s);
+        }
+        // Mostly only so I don't need to implement double ended iterators TODO make this less crap
+        for string in print_vec.iter(){
+            println!("{}", string);
         }
     }
 
+
     pub fn print(&self){
         let stack_iterator = self.into_iter();
+        let mut print_vec: Vec<String> = Vec::new();
+        // Invert order because terminal appears inverted because of push down
         for entry in stack_iterator{
-            println!("{}, {}, {}", entry.rule, entry.start_position, entry.end_position)
+            let s: String = format!("{:?}, {}, {}", entry.rule, entry.start_position, entry.end_position);
+            print_vec.push(s);
+        }
+        for string in print_vec.iter(){
+            println!("{}", string);
         }
     }
 }
@@ -50,7 +69,16 @@ impl<'a> IntoIterator for &'a Stack{
     type Item = &'a StackEntry;
     type IntoIter = StackIter<'a>;
     fn into_iter(self) -> Self::IntoIter{
-        StackIter{stack: &self, count: 0}
+        let vec_size = self.entries.len() as u32;
+        let count = vec_size.checked_sub(1);
+        if count.is_none() {
+            println!("Error: Stack has no contents");
+            exit(-1);
+        }
+        let count = count.unwrap() as u32;
+        StackIter{stack: &self, count: count as usize}
+        //StackIter{stack: &self, count: 0}
+
     }
 }
 
@@ -63,7 +91,11 @@ impl<'a>  Iterator for StackIter<'a>{
     type Item = &'a StackEntry;
     fn next(&mut self) -> Option<&'a StackEntry>{
         let entry = self.stack.read(self.count);
-        self.count += 1;
+        if self.count == 0 {
+            return None
+        }
+        self.count -= 1;
+        //self.count += 1;
         return entry;
     }
 }
@@ -78,11 +110,11 @@ mod tests {
     #[test]
     fn test_stack() {
         // Simulating what would happen after a rule that consumes one character parses and returns true
-        let se = StackEntry{rule: 0, start_position: 5, end_position: 10};
+        let se = StackEntry{rule: Rules::AlphabetLower, start_position: 5, end_position: 10};
         let mut s = Stack::new(100, 20);
         s.push(se);
         let r = s.pop().unwrap();
-        assert_eq!(r.rule, 0);
+        assert_eq!(r.rule as u32, 1);
         assert_eq!(r.start_position, 5);
         assert_eq!(r.end_position, 10);
