+++ parser/src/terminal.rs
